+++ lab01/lab01/Vect.h
@@ -6,20 +6,20 @@
 
 class Vect
 {
-	double _x, _y, _z;
+	float _x, _y, _z;
 public:
 	// Basic Constructor
 	Vect();
 	// General Construction Function
-	Vect(double x, double y, double z);
+	Vect(float x, float y, float z);
 	// destructor
 	~Vect();
 
 	// METHOD functions
 	// get individual coordinates
-	double getVectX() { return _x; }
-	double getVectY() { return _y; }
-	double getVectZ() { return _z; }
+	float getVectX() { return _x; }
+	float getVectY() { return _y; }
+	float getVectZ() { return _z; }
 
 };
 #endif // !_VECT_H
