+++ lab01/lab01/sphere.cpp
@@ -2,7 +2,7 @@
 
 
 
-Sphere::Sphere(Vect* position, double r, Material * mat) : Obj(mat)
+Sphere::Sphere(Vect* position, float r, Material * mat) : Obj(mat)
 {
 	_position = position;
 	_radius = r;
